+++ docs/changelog.rst
@@ -9,6 +9,7 @@ v0.9.6
 * Added web analytics (umami)
 * Added a data export button
 * Added an account deletion button
+* Added a ``clearcache`` command
 * Added a Render demo deployment
 
 v0.9.5